feat: move sharky decoders to the crate & actualize example
