+++ Cargo.toml
@@ -27,6 +27,7 @@ carbon-mpl-token-metadata-decoder = { path = "decoders/mpl-token-metadata-decode
 carbon-orca-whirlpool-decoder = { path = "decoders/orca-whirlpool-decoder", version = "0.4.0" }
 carbon-pumpfun-decoder = { path = "decoders/pumpfun-decoder", version = "0.4.0" }
 carbon-raydium-amm-v4-decoder = { path = "decoders/raydium-amm-v4-decoder", version = "0.4.0" }
+carbon-sharky-decoder = { path = "decoders/sharky-decoder", version = "0.4.0" }
 carbon-system-program-decoder = { path = "decoders/system-program-decoder", version = "0.4.0" }
 carbon-token-program-decoder = { path = "decoders/token-program-decoder", version = "0.4.0" }
 carbon-memo-program-decoder = { path = "decoders/memo-program-decoder", version = "0.4.0" }
