+++ decoders/sharky-decoder/src/instructions/close_order_book.rs
@@ -14,7 +14,7 @@ pub struct CloseOrderBookInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for CloseOrderBook {
     type ArrangedAccounts = CloseOrderBookInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let order_book = accounts.get(0)?;
