+++ decoders/sharky-decoder/src/instructions/create_order_book.rs
@@ -23,7 +23,7 @@ pub struct CreateOrderBookInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for CreateOrderBook {
     type ArrangedAccounts = CreateOrderBookInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let order_book = accounts.get(0)?;
