+++ decoders/sharky-decoder/src/instructions/create_program_version.rs
@@ -18,7 +18,7 @@ pub struct CreateProgramVersionInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for CreateProgramVersion {
     type ArrangedAccounts = CreateProgramVersionInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let authority = accounts.get(0)?;
