+++ decoders/sharky-decoder/src/instructions/extend_loan_v3.rs
@@ -35,7 +35,7 @@ pub struct ExtendLoanV3InstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for ExtendLoanV3 {
     type ArrangedAccounts = ExtendLoanV3InstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let loan = accounts.get(0)?;
