+++ decoders/sharky-decoder/src/instructions/extend_loan_v3_compressed.rs
@@ -37,7 +37,7 @@ pub struct ExtendLoanV3CompressedInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for ExtendLoanV3Compressed {
     type ArrangedAccounts = ExtendLoanV3CompressedInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let loan = accounts.get(0)?;
