+++ decoders/sharky-decoder/src/instructions/rescind_loan.rs
@@ -20,7 +20,7 @@ pub struct RescindLoanInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for RescindLoan {
     type ArrangedAccounts = RescindLoanInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let loan = accounts.get(0)?;
