+++ decoders/sharky-decoder/src/instructions/take_loan_v3_compressed.rs
@@ -32,7 +32,7 @@ pub struct TakeLoanV3CompressedInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for TakeLoanV3Compressed {
     type ArrangedAccounts = TakeLoanV3CompressedInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let lender = accounts.get(0)?;
