+++ decoders/sharky-decoder/src/instructions/update_order_book.rs
@@ -22,7 +22,7 @@ pub struct UpdateOrderBookInstructionAccounts {
 impl carbon_core::deserialize::ArrangeAccounts for UpdateOrderBook {
     type ArrangedAccounts = UpdateOrderBookInstructionAccounts;
 
-fn arrange_accounts(
+    fn arrange_accounts(
         accounts: Vec<solana_sdk::instruction::AccountMeta>,
     ) -> Option<Self::ArrangedAccounts> {
         let order_book = accounts.get(0)?;
