+++ decoders/sharky-decoder/src/types/loan_offer.rs
