+++ decoders/sharky-decoder/src/types/loan_state.rs
