+++ decoders/sharky-decoder/src/types/loan_terms.rs
