+++ decoders/sharky-decoder/src/types/loan_terms_spec.rs
