+++ decoders/sharky-decoder/src/types/taken_loan.rs
