+++ examples/sharky-offers/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 carbon-core = { workspace = true }
 carbon-rpc-program-subscribe-datasource = { workspace = true }
 carbon-log-metrics = { workspace = true }
+carbon-sharky-decoder = { workspace = true }
 solana-client.workspace = true
 solana-sdk.workspace = true
 solana-account-decoder.workspace = true
