--- examples/sharky-offers/src/sharky_decoder/accounts/escrow_pda.rs
@@ -1,11 +1,0 @@
-
- 
-use carbon_core::{borsh, CarbonDeserialize};
-
-#[derive(CarbonDeserialize, Debug)] 
- 
-
-#[carbon(discriminator = "0xa8d91e0ef8c57680")] 
-pub struct EscrowPda { 
-        pub bump: u8, 
-}
