--- examples/sharky-offers/src/sharky_decoder/accounts/nft_list.rs
@@ -1,12 +1,0 @@
-
- 
-use carbon_core::{borsh, CarbonDeserialize};
-
-#[derive(CarbonDeserialize, Debug)] 
- 
-
-#[carbon(discriminator = "0x7c8627b40b49e869")] 
-pub struct NftList { 
-        pub version: u8, 
-        pub collection_name: String, 
-}
